Move cardSum expression inside startgame function
It will not work if 'cardSum' expression is placed outside the function.

First and second cards variables with random numbers are placed inside the 'startGame()' function.

If the 'cardsSum' expression placed outside it will use the first and second card variables outside the 'startGame()' function and they odn't have any value assigned to them.